+++ examples/03-ui-components/04-side-menu-buttons/src/RemoveBlockButton.tsx
@@ -1,15 +1,25 @@
+import {} from "@blocknote/core";
+import { SideMenuExtension } from "@blocknote/core/extensions";
 import {
-  SideMenuProps,
   useBlockNoteEditor,
   useComponentsContext,
+  useExtensionState,
 } from "@blocknote/react";
 import { MdDelete } from "react-icons/md";
 
 // Custom Side Menu button to remove the hovered block.
-export function RemoveBlockButton(props: SideMenuProps) {
+export function RemoveBlockButton() {
   const editor = useBlockNoteEditor();
 
   const Components = useComponentsContext()!;
+
+  const block = useExtensionState(SideMenuExtension, {
+    selector: (state) => state?.block,
+  });
+
+  if (!block) {
+    return null;
+  }
 
   return (
     <Components.SideMenu.Button
@@ -18,7 +28,7 @@ export function RemoveBlockButton(props: SideMenuProps) {
         <MdDelete
           size={24}
           onClick={() => {
-            editor.removeBlocks([props.block]);
+            editor.removeBlocks([block]);
           }}
         />
       }