+++ packages/react/src/components/FormattingToolbar/DefaultButtons/FileDeleteButton.tsx
@@ -4,12 +4,12 @@ import {
   InlineContentSchema,
   StyleSchema,
 } from "@blocknote/core";
-import { useCallback, useMemo } from "react";
+import { useCallback } from "react";
 import { RiDeleteBin7Line } from "react-icons/ri";
 
 import { useComponentsContext } from "../../../editor/ComponentsContext.js";
 import { useBlockNoteEditor } from "../../../hooks/useBlockNoteEditor.js";
-import { useSelectedBlocks } from "../../../hooks/useSelectedBlocks.js";
+import { useEditorState } from "../../../hooks/useEditorState.js";
 import { useDictionary } from "../../../i18n/dictionary.js";
 
 export const FileDeleteButton = () => {
@@ -22,29 +22,43 @@ export const FileDeleteButton = () => {
     StyleSchema
   >();
 
-  const selectedBlocks = useSelectedBlocks(editor);
+  const block = useEditorState({
+    editor,
+    selector: ({ editor }) => {
+      if (!editor.isEditable) {
+        return undefined;
+      }
 
-  const fileBlock = useMemo(() => {
-    // Checks if only one block is selected.
-    if (selectedBlocks.length !== 1) {
-      return undefined;
-    }
+      const selectedBlocks = editor.getSelection()?.blocks || [
+        editor.getTextCursorPosition().block,
+      ];
 
-    const block = selectedBlocks[0];
+      if (selectedBlocks.length !== 1) {
+        return undefined;
+      }
 
-    if (blockHasType(block, editor, block.type, { url: "string" })) {
+      const block = selectedBlocks[0];
+
+      if (
+        !blockHasType(block, editor, block.type, {
+          url: "string",
+        })
+      ) {
+        return undefined;
+      }
+
       return block;
-    }
-
-    return undefined;
-  }, [editor, selectedBlocks]);
+    },
+  });
 
   const onClick = useCallback(() => {
-    editor.focus();
-    editor.removeBlocks([fileBlock!]);
-  }, [editor, fileBlock]);
+    if (block !== undefined) {
+      editor.focus();
+      editor.removeBlocks([block.id]);
+    }
+  }, [block, editor]);
 
-  if (!fileBlock || fileBlock.props.url === "" || !editor.isEditable) {
+  if (block === undefined) {
     return null;
   }
 
@@ -52,11 +66,11 @@ export const FileDeleteButton = () => {
     <Components.FormattingToolbar.Button
       className={"bn-button"}
       label={
-        dict.formatting_toolbar.file_delete.tooltip[fileBlock.type] ||
+        dict.formatting_toolbar.file_delete.tooltip[block.type] ||
         dict.formatting_toolbar.file_delete.tooltip["file"]
       }
       mainTooltip={
-        dict.formatting_toolbar.file_delete.tooltip[fileBlock.type] ||
+        dict.formatting_toolbar.file_delete.tooltip[block.type] ||
         dict.formatting_toolbar.file_delete.tooltip["file"]
       }
       icon={<RiDeleteBin7Line />}