+++ packages/react/src/components/Popovers/FloatingUIOptions.ts
@@ -1,0 +1,17 @@
+import {
+  UseDismissProps,
+  UseFloatingOptions,
+  UseHoverProps,
+  UseTransitionStatusProps,
+  UseTransitionStylesProps,
+} from "@floating-ui/react";
+import { HTMLAttributes } from "react";
+
+export type FloatingUIOptions = {
+  useFloatingOptions?: UseFloatingOptions;
+  useTransitionStylesProps?: UseTransitionStylesProps;
+  useTransitionStatusProps?: UseTransitionStatusProps;
+  useDismissProps?: UseDismissProps;
+  useHoverProps?: UseHoverProps;
+  elementProps?: HTMLAttributes<HTMLDivElement>;
+};