+++ packages/react/src/components/TableHandles/TableHandleMenu/DefaultButtons/TableHeaderButton.tsx
@@ -5,18 +5,22 @@ import {
   InlineContentSchema,
   StyleSchema,
 } from "@blocknote/core";
+import { TableHandlesExtension } from "@blocknote/core/extensions";
 
 import { useComponentsContext } from "../../../../editor/ComponentsContext.js";
 import { useBlockNoteEditor } from "../../../../hooks/useBlockNoteEditor.js";
-import { TableHandleMenuProps } from "../TableHandleMenuProps.js";
+import {
+  useExtension,
+  useExtensionState,
+} from "../../../../hooks/useExtension.js";
 import { useDictionary } from "../../../../i18n/dictionary.js";
 
 export const TableHeaderRowButton = <
   I extends InlineContentSchema = DefaultInlineContentSchema,
   S extends StyleSchema = DefaultStyleSchema,
->(
-  props: TableHandleMenuProps<I, S> & { orientation: "row" | "column" },
-) => {
+>(props: {
+  orientation: "row" | "column";
+}) => {
   const Components = useComponentsContext()!;
   const dict = useDictionary();
 
@@ -25,11 +29,18 @@ export const TableHeaderRowButton = <
     I,
     S
   >();
-  const tableHandles = editor.tableHandles;
+  const tableHandles = useExtension(TableHandlesExtension);
+  const { block, index } = useExtensionState(TableHandlesExtension, {
+    selector: (state) => ({
+      block: state?.block,
+      index: props.orientation === "column" ? state?.colIndex : state?.rowIndex,
+    }),
+  });
 
   if (
-    !tableHandles ||
-    props.index !== 0 ||
+    tableHandles === undefined ||
+    block === undefined ||
+    index !== 0 ||
     props.orientation !== "row" ||
     !editor.settings.tables.headers
   ) {
@@ -37,24 +48,19 @@ export const TableHeaderRowButton = <
   }
 
   // We only support 1 header row for now
-  const isHeaderRow = Boolean(props.block.content.headerRows);
+  const isHeaderRow = Boolean(block.content.headerRows);
 
   return (
     <Components.Generic.Menu.Item
       className={"bn-menu-item"}
       checked={isHeaderRow}
       onClick={() => {
-        // The block may have been modified and out of date, so we get the latest block
-        const block = editor.getBlock(props.block.id);
-        if (!block) {
-          return;
-        }
         editor.updateBlock(block, {
           ...block,
           content: {
             ...block.content,
             headerRows: isHeaderRow ? undefined : 1,
-          },
+          } as any,
         });
       }}
     >
@@ -66,9 +72,9 @@ export const TableHeaderRowButton = <
 export const TableHeaderColumnButton = <
   I extends InlineContentSchema = DefaultInlineContentSchema,
   S extends StyleSchema = DefaultStyleSchema,
->(
-  props: TableHandleMenuProps<I, S> & { orientation: "row" | "column" },
-) => {
+>(props: {
+  orientation: "row" | "column";
+}) => {
   const Components = useComponentsContext()!;
   const dict = useDictionary();
 
@@ -77,11 +83,19 @@ export const TableHeaderColumnButton = <
     I,
     S
   >();
-  const tableHandles = editor.tableHandles;
+  const tableHandles = useExtension(TableHandlesExtension);
+  const block = useExtensionState(TableHandlesExtension, {
+    selector: (state) => state?.block,
+  });
+  const index = useExtensionState(TableHandlesExtension, {
+    selector: (state) =>
+      props.orientation === "column" ? state?.colIndex : state?.rowIndex,
+  });
 
   if (
     !tableHandles ||
-    props.index !== 0 ||
+    index !== 0 ||
+    !block ||
     props.orientation !== "column" ||
     !editor.settings.tables.headers
   ) {
@@ -89,24 +103,19 @@ export const TableHeaderColumnButton = <
   }
 
   // We only support 1 header column for now
-  const isHeaderColumn = Boolean(props.block.content.headerCols);
+  const isHeaderColumn = Boolean(block.content.headerCols);
 
   return (
     <Components.Generic.Menu.Item
       className={"bn-menu-item"}
       checked={isHeaderColumn}
       onClick={() => {
-        // The block may have been modified and out of date, so we get the latest block
-        const block = editor.getBlock(props.block.id);
-        if (!block) {
-          return;
-        }
         editor.updateBlock(block, {
           ...block,
           content: {
             ...block.content,
             headerCols: isHeaderColumn ? undefined : 1,
-          },
+          } as any,
         });
       }}
     >