+++ packages/xl-ai/src/index.ts
@@ -3,7 +3,6 @@ import "./style.css";
 export * from "./AIExtension.js";
 export * from "./components/AIMenu/AIMenu.js";
 export * from "./components/AIMenu/AIMenuController.js";
-export * from "./components/AIMenu/BlockPositioner.js";
 export * from "./components/AIMenu/getDefaultAIMenuItems.js";
 export * from "./components/AIMenu/PromptSuggestionMenu.js";
 export * from "./components/FormattingToolbar/AIToolbarButton.js";