+++ packages/xl-email-exporter/src/react-email/reactEmailExporter.tsx
@@ -292,6 +292,16 @@ export class ReactEmailExporter<
        * Customize the container element
        */
       container?: React.FC<{ children: React.ReactNode }>;
+      /**
+       * Customize the body styles
+       * @default {
+       *   fontFamily: "'SF Pro Display', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Helvetica Neue', Arial, sans-serif",
+       *   fontSize: "16px",
+       *   lineHeight: "1.5",
+       *   color: "#333",
+       * }
+       */
+      bodyStyles?: CSSProperties;
     },
   ) {
     const transformedBlocks = await this.transformBlocks(blocks);
@@ -304,13 +314,15 @@ export class ReactEmailExporter<
       <Html>
         <Head>{options?.head}</Head>
         <Body
-          style={{
-            fontFamily:
-              "'SF Pro Display', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Helvetica Neue', Arial, sans-serif",
-            fontSize: "16px",
-            lineHeight: "1.5",
-            color: "#333",
-          }}
+          style={
+            options?.bodyStyles ?? {
+              fontFamily:
+                "'SF Pro Display', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Helvetica Neue', Arial, sans-serif",
+              fontSize: "16px",
+              lineHeight: "1.5",
+              color: "#333",
+            }
+          }
         >
           {options?.preview && <Preview>{options.preview}</Preview>}
           <Tailwind>