+++ packages/xl-multi-column/src/extensions/ColumnResize/ColumnResizeExtension.ts
@@ -1,4 +1,5 @@
 import { BlockNoteEditor, getNodeById } from "@blocknote/core";
+import { SideMenuExtension } from "@blocknote/core/extensions";
 import { Extension } from "@tiptap/core";
 import { Node } from "prosemirror-model";
 import { Plugin, PluginKey, PluginView } from "prosemirror-state";
@@ -189,7 +190,7 @@ class ColumnResizePluginView implements PluginView {
       this.view.state.tr.setMeta(columnResizePluginKey, newState),
     );
 
-    this.editor.sideMenu.freezeMenu();
+    this.editor.getExtension(SideMenuExtension)?.freezeMenu();
   };
 
   // If the plugin isn't in a resize state, we want to update it to either a
@@ -222,7 +223,10 @@ class ColumnResizePluginView implements PluginView {
 
       // Since the resize bar overlaps the side menu, we don't want to show it
       // if the side menu is already open.
-      if (newState.type === "hover" && this.editor.sideMenu.view?.state?.show) {
+      if (
+        newState.type === "hover" &&
+        this.editor.getExtension(SideMenuExtension)?.store.state?.show
+      ) {
         return;
       }
 
@@ -297,7 +301,7 @@ class ColumnResizePluginView implements PluginView {
       this.view.state.tr.setMeta(columnResizePluginKey, newState),
     );
 
-    this.editor.sideMenu.unfreezeMenu();
+    this.editor.getExtension(SideMenuExtension)?.unfreezeMenu();
   };
 
   destroy() {